pyproject.toml
[project]
name = "configclass"
version = "0.1.0"
description = "New design for this project"
authors = [{ name = "hanzalaareeb", email = "hanzalaareb@gmail.com" }]
requires-python = ">=3.14"
readme = "README.md"
license = { text = "MIT" }
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.14",
]
dependencies = [
    "django (>=6.0,<7.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "ruff (>=0.14.8,<0.15.0)",
    "django-ninja>=1.6.2",
    "psycopg2-binary>=2.9.11",
    "dj-database-url>=3.1.2",
    "python-dotenv>=1.2.2",
    "django-oauth-toolkit>=3.2.0",
    "pydantic[email]>=2.12.5",
]

[build-system]
requires = ["uv_build>=0.10.6,<0.11.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "black==24.10.0",
    "ruff==0.14.8",
]

[tool.uv]
package = false
